// src/api/generated/models/getCompaniesCompanyIdRequestsSortBy.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdRequestsSortBy = typeof GetCompaniesCompanyIdRequestsSortBy[keyof typeof GetCompaniesCompanyIdRequestsSortBy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdRequestsSortBy = {
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
} as const;