// src/api/generated/models/getCompaniesCompanyIdRequestsSortOrder.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdRequestsSortOrder = typeof GetCompaniesCompanyIdRequestsSortOrder[keyof typeof GetCompaniesCompanyIdRequestsSortOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdRequestsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;